Move `get_cute_device_str`
